Update App.css
I'm assuming that the problem was the layout of the UI, which doesn't fill up all the edges of the design as it's supposed to. So, I fixed it by changing the max-width of the app layout in App.css. In that same file, I also fixed the width of the player rating design and the rating table to make it smaller while it's centered, just like how it was before the UI design was expanded.